dromajo_main cleanup
move ancillary functions out of main
new files term_io, block_device, network, options
