+++ include/machine.h
@@ -221,15 +221,15 @@ typedef struct VirtMachine {
     bool stf_memrecord_size_in_bits;  // write memory access size in bits instead of bytes
     bool stf_trace_register_state;    // Enable register records, IN PROGRESS
     bool stf_disable_memory_records;  // Do not place memory accesses into the STF
-    bool stf_tracepoints_enabled;     // Enable detection of start and stop tracepoints
-    bool stf_include_tracepoints;     // Include the trace markers in STF, IN PROGRESS
+//    bool stf_tracepoints_enabled;     // Enable detection of start and stop tracepoints
+//    bool stf_include_tracepoints;     // Include the trace markers in STF, IN PROGRESS
     int  stf_highest_priv_mode;       // Highest privilege mode to trace
     bool stf_force_zero_sha;          // Emit 0 for the SHA in debug tests
 
     bool stf_trace_open;              // STF trace is open (tracing has begun)
-    bool stf_include_stop_tracepoint; // FIXME: no longer used
-    bool stf_in_traceable_region;     // FIXME: no longer used
-    bool stf_in_tracepoint_region;    // FIXME: no longer used
+//    bool stf_include_stop_tracepoint; // FIXME: no longer used
+    bool stf_in_traceable_region;
+//    bool stf_in_tracepoint_region;    // FIXME: no longer used
     bool stf_tracing_enabled;         // Tracing is active
     bool stf_is_start_opc;            // detected the START_TRACE opcode
     bool stf_is_stop_opc;             // detected the STOP_TRACE opcode
