+++ include/options.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <cstdio>
+
+extern FILE* dromajo_stderr;
+extern void usage_isa();
+extern void usage_interactive();
+extern void usage(const char *prog, const char *msg);
