+++ tests/condor/src/andestar/addigp.S
@@ -41,8 +41,7 @@ ASTAR_CODE_BEGIN
   ASTAR_TEST_S18_OP(16,  addigp, 0x00002000, 0x00000000, 0x02000 );
   ASTAR_TEST_S18_OP(17,  addigp, 0x00004000, 0x00000000, 0x04000 );
   ASTAR_TEST_S18_OP(18,  addigp, 0x00008000, 0x00000000, 0x08000 );
-#  ASTAR_TEST_S18_OP(19,  addigp, 0x00010000, 0x00000000, 0x10000 ); FAILS
-#  ASTAR_TEST_S18_OP(20,  addigp, 0xFFFFFFFF, 0x00000000, 0x2FFFF );  # -1
+#  ASTAR_TEST_S18_OP(20,  addigp, 0xFFFFFFFF, 0x00000000, -1 );  # -1
 #  ASTAR_TEST_S18_OP(21,  addigp, 0xFFFFFFFE, 0x00000000, 0x2FFFE );
 #  ASTAR_TEST_S18_OP(22,  addigp, 0xFFFFFFFD, 0x00000000, 0x2FFFD );
 #  ASTAR_TEST_S18_OP(23,  addigp, 0xFFFFFFFB, 0x00000000, 0x2FFFB );
